+++ apps/web/src/boot.ts
@@ -1,0 +1,4 @@
+import "./style.css";
+import { initNav } from "./site-nav";
+
+initNav();